katha full stack/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
google-generativeai==0.8.4
requests==2.32.3
reportlab==4.2.2
chardet==5.2.0